[App] Add DerpPipe
A derpy hack to fix workaround an issue with Angular2.

It takes an array of values and returns a new array with the same
values.

The issue occurs when you try to iderate over a property of a data
structure that was updated via an observable. The change detector goes
all wonky and throws errors unless you repackage the array as a new
array.

Until the Angular2 dev team fixes this, DerpPipe is a viable
alternative.